requirements.txt
Flask==2.0.3
flask-restful==0.3.9
Flask-API==3.0.post1
PyJWT==2.3.0
firebase-admin==5.2.0
